// src/interfaces/appInterfaces.tsx
// Generated by https://quicktype.io

export interface LoginData {
  correo: string;
  password: string;
}
// Generated by https://quicktype.io

export interface RegisterResponse {
  usuario: Usuario;
  token: string;
}

export interface Usuario {
  rol: string;
  estado: boolean;
  google: boolean;
  nombre: string;
  correo: string;
  uid: string;
}

export interface RegisterData {
  correo: string;
  password: string;
  nombre: string;
}
export interface LoginResponse {
  usuario: Usuario;
  token: string;
  msg?: string;
}

export interface Usuario {
  rol: string;
  estado: boolean;
  google: boolean;
  nombre: string;
  correo: string;
  uid: string;
  img?: string;
  // productos:userProducts[];
}
// export interface userProducts{

// }
// Generated by https://quicktype.io

export interface ProductResponse {
  total: number;
  productos: Producto[];
}

export interface Producto {
  precio: string;
  _id: string;
  nombre: string;
  categoria: Categoria;
  usuario: Categoria;
  img?: string;
}

// export interface Categoria {
//   _id: string;
//   nombre: string;
// }
// Generated by https://quicktype.io

export interface ResponseCategories {
  total: number;
  categorias: Categoria[];
}

export interface Categoria {
  _id: string;
  nombre: string;
  usuario?: CreatedBy;
}

export interface CreatedBy {
  _id: string;
  nombre: string;
  img?: string;
}
// Generated by https://quicktype.io

export interface GetUsuarios {
  total: number;
  usuarios: Usuarios[];
}

export interface Usuarios {
  rol: string;
  estado: boolean;
  google: boolean;
  nombre: string;
  correo: string;
  img?: string;
  uid: string;
}